chirp.compare: refactor the storage system a bit
Ignore-this: c9796a53ddbaf3d082f2fa501cb04db1

darcs-hash:20120217143648-21d25-65be99d47739cb38eec7ba06bda0dfc3e2fd934e.gz

+++ chirp/compare/sqlite_storage.py
@@ -8,10 +8,18 @@ as a backend.
 Copyright (C) 2011 Daniel Meliza <dan // meliza.org> Created
 2011-10-05
 """
-import os,sqlite3
+import os,sqlite3,glob
+from .base_storage import base_storage as _base_storage
+
+sql_create_filelist = """\
+CREATE TABLE IF NOT EXISTS %s (
+`id` INTEGER NOT NULL,
+`filename` VARCHAR(512) UNIQUE NOT NULL,
+PRIMARY KEY (`id` ASC AUTOINCREMENT))
+"""
 
 sql_create_target = """\
-CREATE TABLE %s (
+CREATE TABLE IF NOT EXISTS %s (
 `id` INTEGER NOT NULL,
 `ref` INTEGER NOT NULL REFERENCES %s(id),
 `tgt` INTEGER NOT NULL REFERENCES %s(id),
@@ -19,62 +27,65 @@ CREATE TABLE %s (
 PRIMARY KEY (`id` ASC AUTOINCREMENT))
 """
 
-class sqlite_storage(object):
+class sqlite_storage(_base_storage):
+    _descr = "sqlite database (LOCATION: database file)"
+    signal_table = "signals"
 
-    def __init__(self, comparator, location):
+    def __init__(self, comparator, location, signals=None, restrict=False, skip=False, **kwargs):
         """
         Initialize the storage object with the location where the data
-        is stored.  This has the form <path>:<table>, where path is
-        the location of a sqlite database, and table is a table with
-        at least the following fields (id, location). Requires access
-        to the object that will be doing the comparisons (comparator),
-        to determine the file extension and the fields of the output.
+        is stored.
+
+        @param location  the path of a sqlite database (created as needed)
+        @param signals   the directory where the signals are located
+        @param restrict  if True, only include signals already stored in the database;
+                         if False, signals will be added to the database
+        @param skip      if True, pairs() will only return uncompleted calculations
         """
+        _base_storage.__init__(self, comparator)
         self.location = location
-        self.file_pattern = "%s" % comparator.file_extension
-        self.compare_stat_fields = comparator.compare_stat_fields
-        self.table_name = "superb_" + type(comparator).__name__
-        self.symmetric = comparator.symmetric
-        self.signals = []
-        self._load_signals()
+        self.table_name = type(comparator).__name__
+        self.restrict = restrict
+        self.skip = skip
+        self._load_signals(signals or '.')
 
-
-    def _load_signals(self):
+    def _load_signals(self, signal_dir):
         """
         Generates a list of input signals by globbing a directory.
         Returns a list of (id, locator) tuples, where the id is an
         integer code and the locator is the path.  Note glob does not
         always return the same ordering of files.
         """
-        path,table = self.location_parts
-        self.connection = sqlite3.connect(path)
-        cursor = self.connection.cursor()
-        cursor.execute("SELECT id,filename FROM %s ORDER BY id" % table)
-        self.signals = [(i,f+self.file_pattern) for i,f in cursor.fetchall()]
+        self.connection = sqlite3.connect(self.location)
+        with self.connection:
+            cursor = self.connection.cursor()
+            cursor.execute(sql_create_filelist % self.signal_table)
+            files = glob.glob(os.path.join(signal_dir, "*" + self.file_pattern))
+            if not self.restrict:
+                rv = cursor.executemany("INSERT OR IGNORE INTO %s (filename) VALUES (?)" % self.signal_table,
+                                        ((os.path.splitext(os.path.split(x)[1])[0],) for x in files))
+            cursor.execute("SELECT id,filename FROM %s ORDER BY id" % self.signal_table)
+            self.signals = [(i,os.path.join(signal_dir,f+self.file_pattern)) for i,f in cursor.fetchall() \
+                                if os.path.join(signal_dir, f+self.file_pattern) in files]
+
 
     def _create_target_table(self, values):
-        signal_table = self.location_parts[1]
         sql_columns = ",\n".join("`%s` %s" % (name,sqlite_type(val)) for \
                                    name,val in zip(self.compare_stat_fields, values[2:]))
-        sql = sql_create_target % (self.table_name, signal_table, signal_table, sql_columns)
-        self.connection.execute("DROP TABLE IF EXISTS %s" % self.table_name)
+        sql = sql_create_target % (self.table_name, self.signal_table, self.signal_table, sql_columns)
+        if not self.skip:
+            self.connection.execute("DROP TABLE IF EXISTS %s" % self.table_name)
         self.connection.execute(sql)
 
-    @property
-    def location_parts(self):
-        """ A tuple of the database path and the source table """
-        try:
-            path,table = self.location.split(":")
-        except (AttributeError, ValueError):
-            raise ValueError, "Invalid database location (path:table)"
-        if not os.path.exists(path):
-            raise ValueError, "Invalid database location (path does not exist)"
-        return path,table
-
-    @property
-    def nsignals(self):
-        return len(self.signals)
-
+    def pairs(self):
+        if self.skip:
+            cur = self.connection.cursor()
+            cur.execute("SELECT ref,tgt FROM %s" % self.table_name)
+            done = [(r,t) for r,t in cur.fetchall()]
+            for k1,k2 in _base_storage.pairs(self):
+                if (k1,k2) not in done: yield k1,k2
+        else:
+            for k1,k2 in _base_storage.pairs(self): yield k1,k2
 
     def output_signals(self, cout=None):
         """
@@ -90,14 +101,13 @@ class sqlite_storage(object):
         """
         # create table using types of first yielded result
         cols = ("ref","tgt") + tuple(self.compare_stat_fields)
-        sql1 =  "INSERT INTO %s (%s) VALUES (%s)" % (self.table_name,
+        sql1 =  "INSERT OR IGNORE INTO %s (%s) VALUES (%s)" % (self.table_name,
                                                     ",".join(cols),
                                                     ",".join("?" for x in cols))
         cols = ("tgt","ref") + tuple(self.compare_stat_fields)
-        sql2 =  "INSERT INTO %s (%s) VALUES (%s)" % (self.table_name,
+        sql2 =  "INSERT OR IGNORE INTO %s (%s) VALUES (%s)" % (self.table_name,
                                                     ",".join(cols),
                                                     ",".join("?" for x in cols))
-
         result = gen.next()
         with self.connection:
             self._create_target_table(result)
@@ -112,10 +122,14 @@ class sqlite_storage(object):
 * Storage parameters:
 ** Location = %s
 ** File pattern = %s
+** Add files to database = %s
+** Skip pairs already in database = %s
 ** Writing to table %s:%s (fields: %s)""" % (self.location, self.file_pattern,
-                                          self.location_parts[0], self.table_name,
-                                          ",".join(self.compare_stat_fields))
+                                             not self.restrict, self.skip,
+                                             self.location, self.table_name,
+                                             ",".join(self.compare_stat_fields))
         return out
+
 
 # register some sqlite converters
 import numpy
